Refactor(LottoFortuneScreen): Balance의 매개변수 타입변경

+++ app/src/main/java/com/geniusjun/lotto/ui/screen/LottoFortuneScreen.kt
@@ -96,7 +96,7 @@ fun TitleSection() {
 }
 
 @Composable
-fun BalanceCard(balance: Int) {
+fun BalanceCard(balance: Long) {
     AppCard(modifier = Modifier.fillMaxWidth()) {
         Text(text = "현재 보유 금액", color = Color.Gray, fontSize = 13.sp)
         Spacer(modifier = Modifier.height(8.dp))
